#S06P12A305-48 Refactor: RequestBookReviewDto 별점 타입 변경

+++ backend/src/main/java/com/ssafy/bjbj/api/bookinfo/dto/RequestBookReviewDto.java
@@ -13,7 +13,7 @@ public class RequestBookReviewDto {
 
     private Long memberSeq;
 
-    private Integer starRating;
+    private Double starRating;
 
     private String summary;
 }
